Convert file prime_sieve.go in utils to primes.go and integrate into problem-12.go
The file primes.go will now house all functions associated with prime numbers. This change was made to allow for the function GetPrimesLessThan to get a list of primes within a given size rather than using the prime sieve for all prime applications. This change was made to help with problem 12.